// code/src/enums/navLinks.ts
export const enum NAV_LINKS {
  DASHBOARD = '/',
  LOGIN = 'login',
  RESET_PASSWORD = 'reset-password',
  REGISTER = 'register',
  MANAGE_ACCOUNTS = 'manage-accounts',
  LICENSE_PLAN = 'license-plan',
  LICENSE_MANAGEMENT = 'license-management',
  LICENSES_ADMINISTRATION = 'licenses-admin',
  REPORTS_AUDITED = 'reports-audited',
  NOT_FOUND = '*'
}